# data/output/openapi.yaml
openapi: 3.1.0
info:
  title: NavigaTUM
  description: |-
    Navigating around TUM with excellence - An API to search for rooms,
    buildings and other places

    NavigaTUM is a tool developed by students for students, to help you get around at [TUM](https://tum.de). Feel free to contribute.

    - [x] Interactive/static maps to look up the position of rooms or buildings
    - [x] Fast and typo-tolerant search
    - [x] Support for different room code formats as well as generic names
    - [x] All functionality is also available via an open and well documented API
    - [x] Automatically update the data from upstream datasources
    - [x] Allow students/staff to easily submit feedback and data patches
    - [x] Generate turn by turn navigation advice for navigating end to end
    - [ ] Generate maps from CAD data sources

    If you'd like to help out or join us in this adventure, we would love to talk to you.
  termsOfService: https://nav.tum.de/en/about/privacy
  contact:
    name: OpenSource @ TUM e.V.
    url: https://tum.dev/
    email: navigatum@tum.de
  license:
    name: GPL v3
    url: https://www.gnu.org/licenses/
  version: 0.0.0
  x-logo:
    altText: null
    backgroundColor: null
    href: https://nav.tum.de
    url: https://raw.githubusercontent.com/TUM-Dev/NavigaTUM/refs/heads/main/webclient/app/assets/logos/navigatum.svg
servers:
  - url: https://nav.tum.de
    description: production
paths:
...
  /api/maps/route:
    get:
      tags:
        - maps
      summary: Routing requests
      description: |-
        **API IS EXPERIMENTAL AND ACTIVELY SUBJECT TO CHANGE**

        The user specifies using provided origin (`from`) and destination (`to`) locations and a transport mode (`route_costing`) to tune their routing between the two locations.
        The costing is fine-tuned by the server side accordingly.

        Internally, this endpoint relies on
        - [Valhalla](https://github.com/valhalla/valhalla) for routing for route calculation
        - our database to resolve ids.

          You will need to look the ids up via [`/api/search`](#tag/locations/operation/search_handler) beforehand.
          **Note:** [`/api/search`](#tag/locations/operation/search_handler) does support both university internal routing and external addressing.

        **In the future (i.e. public transit routing currently is not implemented)**, it will als rely on either
        - [OpenTripPlanner2](https://www.opentripplanner.org/) or
        - [Motis](https://github.com/motis-project/motis)
      operationId: route_handler
      parameters:
        - name: lang
          in: query
          required: false
          schema:
            type: string
            enum:
              - de
              - en
        - name: from
          in: query
          description: Start of the route
          required: true
          schema:
            oneOf:
              - $ref: '#/components/schemas/Coordinate'
                description: |-
                  Either an
                  - external address which was looked up or
                  - the users current location
              - type: string
                description: Our (uni internal) key for location identification
        - name: to
          in: query
          description: Destination of the route
          required: true
          schema:
            oneOf:
              - $ref: '#/components/schemas/Coordinate'
                description: |-
                  Either an
                  - external address which was looked up or
                  - the users current location
              - type: string
                description: Our (uni internal) key for location identification
        - name: route_costing
          in: query
          description: |-
            Transport mode the user wants to use

            If not specified, the default is based on how far the destinations are apart and requested time.
          required: false
          schema:
            oneOf:
              - type: "null"
              - type: string
                description: Transport mode the user wants to use
                enum:
                  - pedestrian
                  - bicycle
                  - motorcycle
                  - car
                  - public_transit
        - name: pedestrian_type
          in: query
          description: Does the user have specific walking restrictions?
          required: false
          schema:
            type: string
            description: Does the user have specific walking needs?
            enum:
              - standard
              - blind
              - wheelchair
        - name: ptw_type
          in: query
          description: Does the user prefer mopeds or motorcycles for powered two-wheeled (ptw)?
          required: false
          schema:
            type: string
            description: Does the user have a moped or motorcycle
            enum:
              - motorcycle
              - moped
        - name: bicycle_type
          in: query
          description: Which kind of bicycle do you ride?
          required: false
          schema:
            type: string
            description: Which kind of bicycle do you ride?
            enum:
              - road
              - hybrid
              - cross
              - mountain
        - name: page_cursor
          in: query
          description: |-
            Cursor position for pagination
            Only avaliable for some costings
          required: false
          schema:
            type:
              - string
              - "null"
        - name: time
          in: query
          description: |-
            Time for the route (ISO 8601 format)
            Used with `arrive_by` to determine if this is departure or arrival time
          required: false
          schema:
            type:
              - string
              - "null"
            format: date-time
        - name: arrive_by
          in: query
          description: Whether the time parameter represents arrival time (true) or departure time (false/not set)
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: '**Routing solution**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingResponse'
        "404":
          description: '**Not found.** The requested location does not exist'
          content:
            text/plain:
              schema:
                type: string
              example: Not found
...
  /api/search:
    get:
      tags:
        - locations
      summary: Search entries
      description: |-
        This endpoint is designed to support search-as-you-type results.

        Instead of simply returning a list, the search results are returned in a way to provide a richer experience by splitting them up into sections. You might not necessarily need to implement all types of sections, or all sections features (if you just want to show a list). The order of sections is a suggested order to display them, but you may change this as you like.

        Some fields support highlighting the query terms and it uses \x19 and \x17 to mark the beginning/end of a highlighted sequence.
        (See [Wikipedia](https://en.wikipedia.org/wiki/C0_and_C1_control_codes#Modified_C0_control_code_sets)).
        Some text-renderers will ignore them, but in case you do not want to use them, you might want to remove them from the responses via empty `pre_highlight` and `post_highlight` query parameters.
      operationId: search_handler
      parameters:
        - name: q
          in: query
          description: |-
            string you want to search for.

            The amounts returned can be controlled using the `limit_*` parameters.
            Use `in`, `usage`, `type`, and `near` query parameters for filtering.
          required: true
          schema:
            type: string
        - name: in
          in: query
          description: |-
            Filter by parent (building, campus, etc.).

            Can be repeated for multiple values (e.g. `&in=garching&in=5304`).
          required: false
          schema:
            type: array
            items:
              type: string
        - name: usage
          in: query
          description: |-
            Filter by usage type (e.g. `wc`, `büro`).

            Can be repeated for multiple values.
          required: false
          schema:
            type: array
            items:
              type: string
        - name: type
          in: query
          description: |-
            Filter by facet.

            Can be repeated for multiple values. Unknown values cause a `400`.
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/FacetFilter'
        - name: near
          in: query
          description: Sort results by distance to a coordinate (`lat,lon`).
          required: false
          schema:
            type: string
        - name: search_addresses
          in: query
          description: |-
            Include adresses in the saerch

            Be aware that Nominatim (which we use to do this search) is really slow (~100ms).
            Only activate this when you really need it.
          required: false
          schema:
            type: boolean
        - name: limit_sites
          in: query
          description: |-
            Maximum number of sites (campus / site / area) to return.

            Clamped to `0`..`1000`.
            If this is a problem for you, please open an issue.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: limit_buildings
          in: query
          description: |-
            Maximum number of buildings to return.

            Clamped to `0`..`1000`.
            If this is a problem for you, please open an issue.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: limit_rooms
          in: query
          description: |-
            Maximum number of rooms to return.

            Clamped to `0`..`1000`.
            If this is an problem for you, please open an issue.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: limit_pois
          in: query
          description: |-
            Maximum number of POIs (points of interest) to return.

            Clamped to `0`..`1000`.
            If this is a problem for you, please open an issue.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: limit_all
          in: query
          description: |-
            Maximum number of results to return.

            Clamped to `1`..`1000`.
            If this is an problem for you, please open an issue.
          required: false
          schema:
            type: integer
            minimum: 0
        - name: pre_highlight
          in: query
          description: |-
            string to include in front of highlighted sequences.

            If this and `post_highlight` are empty, highlighting is disabled.
            For background on the default values, please see [Wikipedia](https://en.wikipedia.org/wiki/C0_and_C1_control_codes#Modified_C0_control_code_sets)).
          required: false
          schema:
            type: string
        - name: post_highlight
          in: query
          description: |-
            string to include after the highlighted sequences.

            If this and `pre_highlight` are empty, highlighting is disabled.
            For background on the default values, please see [Wikipedia](https://en.wikipedia.org/wiki/C0_and_C1_control_codes#Modified_C0_control_code_sets)).
          required: false
          schema:
            type: string
        - name: cropping
          in: query
          description: |-
            How to handle cropping of long building names in `parsed_id`.

            - `crop` (default): crop long names (> 25 chars) with an ellipsis.
            - `full`: never crop; always show full building names.
          required: false
          schema:
            type: string
            description: Controls whether long building names inside `parsed_id` are cropped.
            enum:
              - crop
              - full
        - name: parsed_id
          in: query
          description: |-
            How to format `parsed_id` for rooms.

            - `prefixed` (default): add common building prefixes (e.g. `MW 1801`).
            - `roomfinder`: return room codes in Roomfinder format (`archname@building_id`).
          required: false
          schema:
            type: string
            description: Controls how `parsed_id` is built for room results.
            enum:
              - prefixed
              - roomfinder
      responses:
        "200":
          description: Search entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        "400":
          description: '**Bad Request.** Not all fields in the body are present as defined above'
          content:
            text/plain:
              schema:
                type: string
              example: 'Query deserialize error: invalid digit found in string'
        "404":
          description: '**Not found.** `q` is empty. Since searching for nothing is nonsensical, we dont support this.'
          content:
            text/plain:
              schema:
                type: string
              example: Not found
        "414":
          description: '**URI Too Long.** The uri you are trying to request is unreasonably long. Search querys dont have thousands of chars..'
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    Coordinate:
      type: object
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
          format: double
          description: Latitude
          example: 48.26244490906312
        lon:
          type: number
          format: double
          description: Longitude
          example: 48.26244490906312
...
    FacetFilter:
      type: string
      description: |-
        Allowlisted values for the `?type=` query parameter.

        Modeled as an enum so `serde` rejects unknown values with a 400 instead of
        silently dropping them, and so the `OpenAPI` schema advertises the exact set
        of accepted values.
      enum:
        - site
        - building
        - room
        - poi
...
    MotisRoutingResponse:
      type: object
      required:
        - direct
        - itineraries
        - next_page_cursor
        - previous_page_cursor
      properties:
        direct:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryResponse'
          description: |-
            Direct trips by `WALK`, `BIKE`, `CAR`, etc. without time-dependency.

             The starting time (`arriveBy=false`) / arrival time
             (`arriveBy=true`) is always the queried `time` parameter (set to
             "now" if not set). But all `direct` connections are meant
             to be independent of absolute times.
        itineraries:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryResponse'
          description: list of itineraries
        next_page_cursor:
          type: string
          description: |-
            Use the cursor to get the next page of results.

            Insert the cursor
             into the request and post it to get the next page.
             The next page is a set of itineraries departing AFTER the last
             itinerary in this result.
        previous_page_cursor:
          type: string
          description: |-
            Use the cursor to get the previous page of results. Insert the
             cursor into the request and post it to get the previous page.
             The previous page is a set of itineraries departing BEFORE the first
             itinerary in the result for a depart after search. When using the
             default sort order the previous set of itineraries is inserted
             before the current result.
...
    ResultsSection:
      type: object
      required:
        - facet
        - entries
        - n_visible
        - estimatedTotalHits
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/ResultEntry'
        estimatedTotalHits:
          type: integer
          description: The estimated (not exact) number of hits for that query
          example: 6
          minimum: 0
        facet:
          $ref: '#/components/schemas/ResultFacet'
          description: These indicate the type of item this represents
        n_visible:
          type: integer
          description: |-
            A recommendation how many of the entries should be displayed by default.

            The number is usually from `0`..`5`.
            More results might be displayed when clicking "expand".
          example: 4
          minimum: 0
...
    RoutingResponse:
      oneOf:
        - allOf:
            - $ref: '#/components/schemas/ValhallaRoutingResponse'
            - type: object
              required:
                - router
              properties:
                router:
                  type: string
                  enum:
                    - valhalla
        - allOf:
            - $ref: '#/components/schemas/MotisRoutingResponse'
            - type: object
              required:
                - router
              properties:
                router:
                  type: string
                  enum:
                    - motis
    SearchResponse:
      type: object
      description: Returned search results by this
      required:
        - sections
        - time_ms
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ResultsSection'
        time_ms:
          type: integer
          format: int32
          description: |-
            Time the search took in the server side, not including network delay

            Maximum as timeout.
            other timeouts (browser, your client) may be smaller.
            Expected average is `10`..`50` for uncached, regular requests.
          example: 8
          minimum: 0
...
    SummaryResponse:
      type: object
      required:
        - time_seconds
        - length_meters
        - has_toll
        - has_highway
        - has_ferry
        - min_lat
        - min_lon
        - max_lat
        - max_lon
      properties:
        has_ferry:
          type: boolean
          description: if the path uses one or more ferry segments
        has_highway:
          type: boolean
          description: If the path uses one or more highway segments
        has_toll:
          type: boolean
          description: If the path uses one or more toll segments
        length_meters:
          type: number
          format: double
          description: Distance traveled in meters
          example: 103.01
        max_lat:
          type: number
          format: double
          description: Maximum latitude of the sections bounding box
          example: 48.26244490906312
        max_lon:
          type: number
          format: double
          description: Maximum longitude of the sections bounding box
          example: 48.26244490906312
        min_lat:
          type: number
          format: double
          description: Minimum latitude of the sections bounding box
          example: 48.26244490906312
        min_lon:
          type: number
          format: double
          description: Minimum longitude of the sections bounding box
          example: 48.26244490906312
        time_seconds:
          type: number
          format: double
          description: Estimated elapsed time in seconds
          example: 201.025
...
    ValhallaLegResponse:
      type: object
      required:
        - summary
        - maneuvers
        - shape
      properties:
        maneuvers:
          type: array
          items:
            $ref: '#/components/schemas/ManeuverResponse'
        shape:
          type: array
          items:
            $ref: '#/components/schemas/Coordinate'
        summary:
          $ref: '#/components/schemas/SummaryResponse'
    ValhallaRoutingResponse:
      type: object
      required:
        - legs
        - summary
      properties:
        legs:
          type: array
          items:
            $ref: '#/components/schemas/ValhallaLegResponse'
          description: |-
            A trip contains one (or more) legs.

            A leg is created when routing stops, which currently only happens at the ends (`from`, `to`).
          maxItems: 1
          minItems: 1
        summary:
          $ref: '#/components/schemas/SummaryResponse'
          description: Trip summary
...